moved get_tau_integ
